requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
cmake==4.0.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cycler==0.12.1
Django==4.2.15
django-bootstrap3==25.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-redis==6.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-MySQL==1.6.0
fonttools==4.59.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.5
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
phonenumbers==9.0.24
pillow==11.1.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.4
python-pptx==1.0.2
pytz==2021.1
razorpay==1.4.2
redis==7.4.0
reportlab==4.4.3
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
setuptools==78.1.0
six==1.17.0
sqlparse==0.4.1
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.9
